feat: integrate components in live chat
